adding a description for the driver step
