+++ cmd/template/dbdriver/mongo.go
@@ -1,4 +1,4 @@
-package DBDriver
+package dbdriver
 
 type MongoTemplate struct{}
